reorganize the new code
