+++ packages/roosterjs-editor-dom/lib/deprecated/selectionMarker.ts
@@ -1,5 +1,5 @@
-import createRange from './createRange';
-import Position from './Position';
+import createRange from '../selection/createRange';
+import Position from '../selection/Position';
 import queryElements from '../utils/queryElements';
 import { NodePosition, PositionType } from 'roosterjs-editor-types';
