+++ packages/roosterjs-editor-dom/lib/index.ts
@@ -48,9 +48,9 @@ export { default as VTable, VCell } from './table/VTable';
 
 export { default as Position } from './selection/Position';
 export { default as createRange } from './selection/createRange';
-export { default as getPositionPath } from './selection/getPositionPath';
 export { default as getPositionRect } from './selection/getPositionRect';
 export { default as isPositionAtBeginningOf } from './selection/isPositionAtBeginningOf';
+export { default as getSelectionPath, getRangeFromSelectionPath } from './selection/getSelectionPath';
 
 // Deprecated
 export { default as isTextualInlineElement } from './deprecated/isTextualInlineElement';
@@ -62,4 +62,4 @@ export {
     SanitizeHtmlPropertyCallback,
     StyleMap,
 } from './utils/sanitizeHtml';
-export { markSelection, removeMarker } from './selection/selectionMarker';
+export { markSelection, removeMarker } from './deprecated/selectionMarker';
