+++ packages/roosterjs-editor-types/lib/index.ts
@@ -47,6 +47,7 @@ export { default as InsertOption } from './interface/InsertOption';
 export { default as LinkData } from './interface/LinkData';
 export { default as NodePosition } from './interface/NodePosition';
 export { default as Rect } from './interface/Rect';
+export { default as SelectionPath } from './interface/SelectionPath';
 export { default as TableFormat } from './interface/TableFormat';
 
 // Legacy
