+++ packages/roosterjs-editor-types/lib/interface/SelectionPath.ts
@@ -1,0 +1,14 @@
+/**
+ * A start and end pair of selection path
+ */
+export default interface SelectionPath {
+    /**
+     * Start path of selection
+     */
+    start: number[];
+
+    /**
+     * End path of selection
+     */
+    end: number[];
+}
